Move retry properties to LoadBalancerRetryProperties.
