+++ spring-cloud-commons/src/main/java/org/springframework/cloud/client/loadbalancer/reactive/LoadBalancerProperties.java
@@ -47,19 +47,6 @@ public class LoadBalancerProperties {
 		this.healthCheck = healthCheck;
 	}
 
-	/**
-	 * Properties for LoadBalancer retries.
-	 */
-	private Retry retry = new Retry();
-
-	public Retry getRetry() {
-		return retry;
-	}
-
-	public void setRetry(Retry retry) {
-		this.retry = retry;
-	}
-
 	public static class HealthCheck {
 
 		/**
@@ -98,58 +85,5 @@ public class LoadBalancerProperties {
 			this.interval = interval;
 		}
 
-	}
-
-	public static class Retry {
-
-		/**
-		 * Indicates retries should be attempted on operations other than {@link HttpMethod#GET}.
-		 */
-		private boolean retryOnAllOperations = false;
-
-		/**
-		 * Number of retries to be executed on the same {@link ServiceInstance}.
-		 */
-		private int maxRetriesOnSameServiceInstance = 0;
-
-		/**
-		 * Number of retries to be executed on the next {@link ServiceInstance}.
-		 * A {@link ServiceInstance is chosen before each retry call}.
-		 */
-		private int maxRetriesOnNextServiceInstance = 1;
-
-		private Set<Integer> retryableStatusCodes;
-
-		public boolean isRetryOnAllOperations() {
-			return retryOnAllOperations;
-		}
-
-		public void setRetryOnAllOperations(boolean retryOnAllOperations) {
-			this.retryOnAllOperations = retryOnAllOperations;
-		}
-
-		public int getMaxRetriesOnSameServiceInstance() {
-			return maxRetriesOnSameServiceInstance;
-		}
-
-		public void setMaxRetriesOnSameServiceInstance(int maxRetriesOnSameServiceInstance) {
-			this.maxRetriesOnSameServiceInstance = maxRetriesOnSameServiceInstance;
-		}
-
-		public int getMaxRetriesOnNextServiceInstance() {
-			return maxRetriesOnNextServiceInstance;
-		}
-
-		public void setMaxRetriesOnNextServiceInstance(int maxRetriesOnNextServiceInstance) {
-			this.maxRetriesOnNextServiceInstance = maxRetriesOnNextServiceInstance;
-		}
-
-		public Set<Integer> getRetryableStatusCodes() {
-			return retryableStatusCodes;
-		}
-
-		public void setRetryableStatusCodes(Set<Integer> retryableStatusCodes) {
-			this.retryableStatusCodes = retryableStatusCodes;
-		}
 	}
 }
